backend/flip-logo/requirements.txt
Pillow
requests
boto3
